Created configuration file for the interfaces

+++ game/src/gameWorld.cpp
@@ -24,9 +24,9 @@ GameWorld::GameWorld(int cwidth, int cheight) {
     glBufferData(GL_ARRAY_BUFFER, buffSize * sizeof(GLfloat), nullptr, GL_DYNAMIC_DRAW);
     size_t stride = 5 * sizeof(GLfloat);
     glEnableVertexAttribArray(0); // Center's pos
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, stride, (void*)(0 * sizeof(GLuint)));
+    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, stride, (void*)(0 * sizeof(GLfloat)));
     glEnableVertexAttribArray(1); // Texture coordinates
-    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, stride, (void*)(3 * sizeof(GLuint)));
+    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, stride, (void*)(3 * sizeof(GLfloat)));
 
     glBindBuffer(GL_ARRAY_BUFFER, 0);
     glBindVertexArray(0);
